Philosophical commit.
Just came to the conclusion (?) that tags are not needed: if a node exists which fills your requirement, you should use it. If you are working on a test graph, you may tag the edge as temporary or "test.

This means that there are two kinds of graphs: "PREMIUM" which are authoritative and "TEMPORARY" which are tentative. How do I solve the problem of two ontologies describing the same thing?
With tags?
Preventing it? In this case we need good tools to transfer one branch to another tree.

Tags are needed, because an annotation cannot be explained with a node reference, but only with a sequence of nodes forming a path. Now, should the path be stored as a sequence of Node/predicate/node...Node? Or a sequence of Term/Predicate term/Term...Term?
Or both?
This means that once a node is created, it cannot be deleted. Not an issue for me...
The bottom line is that we need tags.

But what should a tag record?

Certainly the terms, these give us immediately the human definition of the annotation, nodes will provide us an immediate entry into the graph, and an anchor for reasoning.
At this point I guess I can safely develop the node and the edges, leaving to the tags the burden of making all this make sense, at least to any poor person reading these lines...